move helper functions to solvers.py
